Move Scale to own file
